+++ decidim-core/lib/decidim/core/engine.rb
@@ -269,7 +269,7 @@ module Decidim
       end
 
       initializer "Expire sessions" do
-        Rails.application.config.session_store :active_record_store, :key => '_decidim_session', expire_after: Decidim.config.expire_session_after
+        Rails.application.config.session_store :active_record_store, key: "_decidim_session", expire_after: Decidim.config.expire_session_after
       end
 
       initializer "decidim.core.register_resources" do